\name{dmr.attrsel-package}
\alias{dmr.attrsel-package}
\alias{dmr.attrsel}

\docType{package}

\title{A package with example functions from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{A package with example functions from Chapter 19 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\details{See examples in Chapter 19.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\keyword{package}
\keyword{models}
\keyword{multivariate}

\seealso{
\code{\link[dmr.util:dmr.util-package]{dmr.util}}
\code{\link[dmr.stats:dmr.stats-package]{dmr.stats}}
}

\examples{
}
